Generate targets in empty areas.
Method uses multiple heuristics: First we compute a concave hull by
computing a Delaunay triangulation then pruning triangles with long
edges. We then draw the triangles onto a canvas and fill small holes.

The mask is computed globally and then for the floor/ceiling. The masks can
then be intersected (with the floor mask complemented) to obtain
positions that need to be scanned.